Small improvments to some admin commands!

+++ modules/admin.py
@@ -9,24 +9,21 @@ with open("database/adminweenie.json","r") as infile:
 
 '''Clear/Removes a given amount of messages.'''
 async def clear_logic(message, client):
-    if str(message.author.id) in admin:
-        if message.author.permissions_in(message.channel).manage_messages:
-            try:
-                amount = message.content.split(' ')
-                amount_number = amount[1]
-                amount = int(amount_number) + 1
-                print(message.author.name + ' cleared {} messages'.format(amount))
-                deleted = await client.purge_from(message.channel, limit=int(amount), check=None)
-                tbd = await client.send_message(message.channel, 'Deleted {} message(s)'.format(len(deleted)))
-                await asyncio.sleep(5)
-                await client.delete_message(tbd)
-            except ValueError:
-                await client.send_message(message.channel, 'Error, Did you specify number?')
-        else:
-            await client.send_message(message.channel, 'You need manage messages permission on this server, to use this commands')
-    elif str(message.author.id) not in admin:
-        await client.send_message(message.channel, 'Only Admins can Clear channels! If you would like to get admin please contact ' + client.bot_info.owner.id)
-
+    if message.author.permissions_in(message.channel).manage_messages:
+        try:
+            amount = message.content.split(' ')
+            amount_number = amount[1]
+            amount = int(amount_number) + 1
+            print(message.author.name + ' cleared {} messages'.format(amount))
+            deleted = await client.purge_from(message.channel, limit=int(amount), check=None)
+            await asyncio.sleep(1)
+            tbd = await client.send_message(message.channel, 'Deleted {} message(s)'.format(len(deleted)))
+            await asyncio.sleep(8)
+            await client.delete_message(tbd)
+        except ValueError:
+            await client.send_message(message.channel, 'Error, Did you specify number?')
+    else:
+        await client.send_message(message.channel, 'You need manage messages permission on this server, to use this commands')
 commands.add_command(command_name='clear', command_function=clear_logic)
 
 async def admintest_logic(message, client):
@@ -153,8 +150,9 @@ commands.add_command(command_name='admins', command_function=admin_amount, alias
 
 '''Purges channel of ALL messages!.'''
 async def purge(message, client):
-    if str(message.author.id) in admin:
+    if str(message.author.id) in admin and message.author.permissions_in(message.channel).manage_messages:
         deleted = await client.purge_from(message.channel, limit=500, check=None)
+        await asyncio.sleep(1)
         await client.send_message(message.channel, 'Deleted {} message(s)'.format(len(deleted)))
     elif str(message.author.id) not in admin:
         await client.send_message(message.channel, 'Only Admins can Purge channels!. If you would like to get admin please contact ' + client.bot_info.owner.id)
